refactor packages and service's methods
