+++ src/main/java/com/rebecamontag/projectuniversity/model/entity/Professor.java
@@ -1,15 +1,13 @@
-package com.rebecamontag.projectuniversity.entity;
+package com.rebecamontag.projectuniversity.model.entity;
 
-import com.rebecamontag.projectuniversity.enumeration.Gender;
+import com.rebecamontag.projectuniversity.model.enumeration.Gender;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.SequenceGenerator;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -30,11 +28,11 @@ import java.util.Objects;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Student implements Serializable {
+public class Professor implements Serializable {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "student_sequence")
-    @SequenceGenerator(name = "student_sequence", sequenceName = "std_seq", allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "professor_sequence")
+    @SequenceGenerator(name = "professor_sequence", sequenceName = "prof_seq", allocationSize = 1)
     private Integer id;
 
     private String firstName;
@@ -50,10 +48,7 @@ public class Student implements Serializable {
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
-    @ManyToMany
-    @JoinTable(name = "registration",
-                joinColumns = @JoinColumn(name = "student_id"),
-                inverseJoinColumns = @JoinColumn(name = "course_id"))
+    @OneToMany(mappedBy = "professor")
     @ToString.Exclude
     private List<Course> courses;
 
@@ -62,8 +57,8 @@ public class Student implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Student student = (Student) o;
-        return id.equals(student.id);
+        Professor professor = (Professor) o;
+        return id.equals(professor.id);
     }
 
     @Override
