+++ src/main/java/com/rebecamontag/projectuniversity/model/entity/Student.java
@@ -1,13 +1,15 @@
-package com.rebecamontag.projectuniversity.entity;
+package com.rebecamontag.projectuniversity.model.entity;
 
-import com.rebecamontag.projectuniversity.enumeration.Gender;
+import com.rebecamontag.projectuniversity.model.enumeration.Gender;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
 import jakarta.persistence.SequenceGenerator;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -28,11 +30,11 @@ import java.util.Objects;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Professor implements Serializable {
+public class Student implements Serializable {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "professor_sequence")
-    @SequenceGenerator(name = "professor_sequence", sequenceName = "prof_seq", allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "student_sequence")
+    @SequenceGenerator(name = "student_sequence", sequenceName = "std_seq", allocationSize = 1)
     private Integer id;
 
     private String firstName;
@@ -48,7 +50,10 @@ public class Professor implements Serializable {
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
-    @OneToMany(mappedBy = "professor")
+    @ManyToMany
+    @JoinTable(name = "registration",
+                joinColumns = @JoinColumn(name = "student_id"),
+                inverseJoinColumns = @JoinColumn(name = "course_id"))
     @ToString.Exclude
     private List<Course> courses;
 
@@ -57,8 +62,8 @@ public class Professor implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Professor professor = (Professor) o;
-        return id.equals(professor.id);
+        Student student = (Student) o;
+        return id.equals(student.id);
     }
 
     @Override
